Added Base32 decoded data for secret

+++ SimplyAuth/SimplyAuth/SimplyAuthApp.swift
@@ -16,7 +16,7 @@ struct SimplyAuthApp: App {
             store: Store(
               initialState: HomeState(
                 passwords: [],
-                unixEpochSeconds: UInt(Date().timeIntervalSince1970)
+                unixEpochSeconds: UInt64(Date().timeIntervalSince1970)
               ),
               reducer: homeReducer,
               environment: .live
